Move monitor-related code into monitor.go
